# spec/openapi.yaml
openapi: 3.1.0
info:
  title: Aster DNS API
  version: 1.0.0
  description: |
    REST API for the self-hosted Aster DNS control plane. DNS Providers are the
    authoritative source for zones and records. Cookie-authenticated mutations
    require both a trusted Origin and the X-CSRF-Token header.
servers:
  - url: /
...
paths:
...
  /api/v1/auth/bootstrap:
    get:
      tags: [Authentication]
      summary: Read first-administrator bootstrap status
      operationId: getBootstrapStatus
      responses:
        '200':
          description: Bootstrap status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BootstrapStatus' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/bootstrap/passkey/options:
    post:
      tags: [Authentication]
      summary: Begin first-administrator Passkey registration
      operationId: beginBootstrapPasskey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [bootstrap_token, username, display_name, passkey_name]
              properties:
                bootstrap_token: { type: string, writeOnly: true }
                username: { type: string }
                display_name: { type: string }
                passkey_name: { type: string }
      responses:
        '200': { $ref: '#/components/responses/RegistrationOptions' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/bootstrap/passkey/verify:
    post:
      tags: [Authentication]
      summary: Finish first-administrator Passkey registration
      operationId: finishBootstrapPasskey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/CeremonyVerification'
                - type: object
                  required: [bootstrap_token]
                  properties:
                    bootstrap_token: { type: string, writeOnly: true }
      responses:
        '201': { $ref: '#/components/responses/Login' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/bootstrap/password:
    post:
      tags: [Authentication]
      summary: Create the first administrator with a password
      operationId: bootstrapPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [bootstrap_token, username, display_name, password]
              properties:
                bootstrap_token: { type: string, writeOnly: true }
                username: { type: string }
                display_name: { type: string }
                password: { type: string, format: password, writeOnly: true }
      responses:
        '201': { $ref: '#/components/responses/Login' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/passkeys/login/options:
    post:
      tags: [Authentication]
      summary: Begin Passkey login
      operationId: beginPasskeyLogin
      requestBody: { $ref: '#/components/requestBodies/EmptyJSON' }
      responses:
        '200': { $ref: '#/components/responses/AuthenticationOptions' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/passkeys/login/verify:
    post:
      tags: [Authentication]
      summary: Finish Passkey login
      operationId: finishPasskeyLogin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CeremonyVerification' }
      responses:
        '200': { $ref: '#/components/responses/Login' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/passkeys/enroll/options:
    post:
      tags: [Authentication]
      summary: Begin one-time-token Passkey enrollment
      operationId: beginPasskeyEnrollment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [enrollment_token, passkey_name]
              properties:
                enrollment_token: { type: string, writeOnly: true }
                passkey_name: { type: string }
      responses:
        '200': { $ref: '#/components/responses/RegistrationOptions' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/passkeys/enroll/verify:
    post:
      tags: [Authentication]
      summary: Finish one-time-token Passkey enrollment
      operationId: finishPasskeyEnrollment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/CeremonyVerification'
                - type: object
                  required: [enrollment_token]
                  properties:
                    enrollment_token: { type: string, writeOnly: true }
      responses:
        '201': { $ref: '#/components/responses/Login' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/login/password:
    post:
      tags: [Authentication]
      summary: Log in with password fallback
      operationId: passwordLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password, writeOnly: true }
      responses:
        '200': { $ref: '#/components/responses/Login' }
        '202':
          description: Password verified; TOTP completion is required.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/login/totp:
    post:
      tags: [Authentication]
      summary: Complete password login with TOTP
      operationId: totpLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [totp_token, code]
              properties:
                totp_token: { type: string, writeOnly: true }
                code: { type: string, pattern: '^[0-9]{6}$', writeOnly: true }
      responses:
        '200': { $ref: '#/components/responses/Login' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/session:
    get:
      tags: [Authentication]
      summary: Read current session
      operationId: getCurrentSession
      security: [{ sessionCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/Session' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/logout:
    post:
      tags: [Authentication]
      summary: Log out current session
      operationId: logout
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CSRFToken' }]
      requestBody: { $ref: '#/components/requestBodies/EmptyJSON' }
      responses:
        '200': { $ref: '#/components/responses/Status' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/logout-all:
    post:
      tags: [Authentication]
      summary: Revoke every user session
      operationId: logoutAll
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CSRFToken' }]
      requestBody: { $ref: '#/components/requestBodies/EmptyJSON' }
      responses:
        '200': { $ref: '#/components/responses/Status' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/passkeys:
    get:
      tags: [Authentication]
      summary: List safe Passkey metadata
      operationId: listPasskeys
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Passkeys
          content:
            application/json:
              schema:
                type: object
                required: [passkeys]
                properties:
                  passkeys:
                    type: array
                    items: { $ref: '#/components/schemas/Passkey' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/passkeys/register/options:
    post:
      tags: [Authentication]
      summary: Begin registration of another Passkey
      operationId: beginPasskeyRegistration
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CSRFToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties:
                name: { type: string }
      responses:
        '200': { $ref: '#/components/responses/RegistrationOptions' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/passkeys/register/verify:
    post:
      tags: [Authentication]
      summary: Finish registration of another Passkey
      operationId: finishPasskeyRegistration
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CSRFToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CeremonyVerification' }
      responses:
        '201': { $ref: '#/components/responses/Login' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/passkeys/{id}:
    delete:
      tags: [Authentication]
      summary: Delete a Passkey
      operationId: deletePasskey
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/ResourceID'
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '200': { $ref: '#/components/responses/Login' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/password:
    put:
      tags: [Authentication]
      summary: Enable or replace password fallback
      operationId: setPassword
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CSRFToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [password]
              properties:
                password: { type: string, minLength: 12, maxLength: 1024, writeOnly: true }
      responses:
        '200': { $ref: '#/components/responses/Login' }
        default: { $ref: '#/components/responses/Error' }
    delete:
      tags: [Authentication]
      summary: Disable password fallback
      operationId: deletePassword
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CSRFToken' }]
      responses:
        '200': { $ref: '#/components/responses/Login' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/totp/setup:
    post:
      tags: [Authentication]
      summary: Create a pending TOTP enrollment
      operationId: setupTOTP
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CSRFToken' }]
      requestBody: { $ref: '#/components/requestBodies/EmptyJSON' }
      responses:
        '200':
          description: One-time provisioning URI
          content:
            application/json:
              schema:
                type: object
                required: [provisioning_uri]
                properties:
                  provisioning_uri: { type: string, description: Returned once for the pending enrollment. }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/totp/confirm:
    post:
      tags: [Authentication]
      summary: Confirm and enable TOTP
      operationId: confirmTOTP
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CSRFToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [code]
              properties:
                code: { type: string, pattern: '^[0-9]{6}$', writeOnly: true }
      responses:
        '200': { $ref: '#/components/responses/Login' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/totp:
    delete:
      tags: [Authentication]
      summary: Disable TOTP
      operationId: deleteTOTP
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CSRFToken' }]
      responses:
        '200': { $ref: '#/components/responses/Login' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/sessions:
    get:
      tags: [Authentication]
      summary: List active server-side sessions
      operationId: listSessions
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Active sessions
          content:
            application/json:
              schema:
                type: object
                required: [sessions]
                properties:
                  sessions:
                    type: array
                    items: { $ref: '#/components/schemas/SessionInfo' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/sessions/revoke-others:
    post:
      tags: [Authentication]
      summary: Revoke all sessions except the current session
      operationId: revokeOtherSessions
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CSRFToken' }]
      requestBody: { $ref: '#/components/requestBodies/EmptyJSON' }
      responses:
        '200': { $ref: '#/components/responses/Status' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/sessions/{id}:
    delete:
      tags: [Authentication]
      summary: Revoke another session
      operationId: revokeSession
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/ResourceID'
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '200': { $ref: '#/components/responses/Status' }
        default: { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-aster_session
      description: Opaque session cookie. HTTP deployments use the non-Host-prefixed equivalent.
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
      description: Must match the readable CSRF cookie for the current server-side session.
    ResourceID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  requestBodies:
    EmptyJSON:
      required: false
      content:
        application/json:
          schema: { type: object, maxProperties: 0 }
  responses:
    Error:
      description: Stable safe error. Details never contain Provider secrets or raw upstream payloads.
      headers:
        X-Request-ID:
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Status:
      description: Operation status
      content:
        application/json:
          schema:
            type: object
            required: [status]
            properties:
              status: { type: string }
    RegistrationOptions:
      description: WebAuthn registration options and one-time ceremony token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/CeremonyOptions' }
    AuthenticationOptions:
      description: WebAuthn authentication options and one-time ceremony token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/CeremonyOptions' }
    Login:
      description: Authentication transition result
      content:
        application/json:
          schema: { $ref: '#/components/schemas/LoginResponse' }
    Session:
      description: Current authenticated session
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SessionResponse' }
...
  schemas:
...
    BootstrapStatus:
      type: object
      required: [required, configured, password_login_enabled]
      properties:
        required: { type: boolean }
        configured: { type: boolean }
        password_login_enabled: { type: boolean }
...
    CeremonyVerification:
      type: object
      additionalProperties: false
      required: [ceremony_token, credential]
      properties:
        ceremony_token: { type: string, writeOnly: true }
        credential: { type: object, additionalProperties: true }
...
    LoginResponse:
      type: object
      required: [authenticated, totp_required]
      properties:
        authenticated: { type: boolean }
        user: { $ref: '#/components/schemas/User' }
        totp_required: { type: boolean }
        totp_token: { type: string, description: Short-lived token returned when password login requires TOTP. }
...